+++ Module.php
@@ -14,6 +14,10 @@ use yii\helpers\Url;
  */
 class Module extends \humhub\components\Module
 {
+    /**
+     * @inheritdoc
+     */
+    public $resourcesPath = 'assets';
 
     /**      *
      * @var int the sort order for the birthdays sidebar widget