+++ assets/Assets.php
@@ -1,22 +1,18 @@
 <?php
-
 /**
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2015 HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
  */
 
-namespace humhub\modules\birthday;
+namespace humhub\modules\birthday\assets;
 
 use yii\web\AssetBundle;
 
 class Assets extends AssetBundle
 {
-
-    public function init()
-    {
-        $this->sourcePath = dirname(__FILE__) . '/assets';
-        parent::init();
-    }
-
+    /**
+     * @inheritdoc
+     */
+    public $sourcePath = '@birthday/resources';
 }